.infra/requirements.txt
aws-cdk.aws-events==1.138.2
aws-cdk.aws-events-targets==1.138.2
aws-cdk.aws-lambda-python==1.138.2
aws-cdk.core==1.138.2
aws-cdk.pipelines==1.138.2
aws-cdk.aws-codepipeline==1.138.2
aws-cdk.aws-codepipeline-actions==1.138.2
aws-cdk.cx-api==1.138.2
aws-cdk.cloud-assembly-schema==1.138.2
aws_cdk.aws_ec2==1.138.2
aws_cdk.aws_ecs==1.138.2
aws_cdk.aws_ecs_patterns==1.138.2
requests
boto3
cdk_ecr_deployment
